Update README.md
After finishing my internship at Research and Innovative Center in Near east University, I'm happy to share with you the project  i have done during this internship with the help of my great Supervisors. this project aims is to operates through a combination of components that work together to create an interactive and automated lighting system. At the heart of the system is the Arduino board, which serves as the central controller. It processes inputs from the Voice recognition module, which allows users to issue voice commands to control the system. when a voice command is detected, the Arduino interprets the command and sends signals to the Relay Module. These Relay module act as electronic switches that control the power supply to the connected devices, such as Fan and the Light Bulb.